LayoutTests/fast/events/popup-blocked-from-fake-focus-expected.txt


PASS win is undefined.
PASS win is undefined.
